PoiDB submission (#503)

+++ ElectronicObserver/Data/TsunDbSubmission/Battle/TsunDbFleetsAndAirBaseData.cs
@@ -56,7 +56,7 @@ public class TsunDbFleetsAndAirBaseData : TsunDbEntity
 			.Min();
 
 		// --- Get the fleet type, if first fleet => flag of the combined fleet, else 0 (single fleet & strike force)
-		FleetType = sortiedFleet == 1 ? db.Fleet.CombinedFlag : 0;
+		FleetType = sortiedFleet == 1 ? (int)db.Fleet.CombinedFlag : 0;
 
 		// --- Fleet 1
 		Fleet1 = PrepareFleet(db.Fleet[sortiedFleet]);
